- Duke: eliminated all local short variables in actors*.cpp.

+++ source/games/duke/src/duke3d.h
@@ -92,10 +92,10 @@ struct Dispatcher
 	bool (*floorspace)(sectortype* sectp);
 	void (*addweapon)(struct player_struct *p, int weapon);
 	void (*hitradius)(DDukeActor* i, int  r, int  hp1, int  hp2, int  hp3, int  hp4);
-	void (*lotsofmoney)(DDukeActor *s, short n);
-	void (*lotsofmail)(DDukeActor *s, short n);
-	void (*lotsofpaper)(DDukeActor *s, short n);
-	void (*guts)(DDukeActor* s, short gtype, short n, short p);
+	void (*lotsofmoney)(DDukeActor *s, int n);
+	void (*lotsofmail)(DDukeActor *s, int n);
+	void (*lotsofpaper)(DDukeActor *s, int n);
+	void (*guts)(DDukeActor* s, int gtype, int n, int p);
 	DDukeActor* (*ifhitsectors)(int sectnum);
 	int  (*ifhitbyweapon)(DDukeActor* sectnum);
 	void (*fall)(DDukeActor* actor, int g_p);
